feat: add redux with thunk to get product detail
