+++ redux/actions/counter.js
@@ -1,0 +1,11 @@
+export const increment = () => {
+  return ({
+    type: 'COUNTER_INCREMENT'
+  })
+}
+
+export const decrement = () => {
+  return ({
+    type: 'COUNTER_DECREMENT'
+  })
+}
