+++ redux/reducers/counter.js
@@ -1,0 +1,25 @@
+const initialState = {
+  num: 1
+}
+
+const counter = (state = initialState, action) => {
+  switch (action.type) {
+    case 'COUNTER_INCREMENT': {
+      state.num = state.num + 1
+      return { ...state }
+    }
+    case 'COUNTER_DECREMENT': {
+      if (state.num > 0) {
+        state.num = state.num - 1
+        return { ...state }
+      } else {
+        return { ...state }
+      }
+    }
+    default: {
+      return { ...state }
+    }
+  }
+}
+
+export default counter
