+++ redux/reducers/index.js
@@ -1,8 +1,14 @@
 import { combineReducers } from "redux";
 import auth from "./auth";
+import productDetail from "./productDetail";
+import counter from "./counter";
+import pages from "./pages";
 
 const rootReducers = combineReducers({
-  auth
+  auth,
+  productDetail,
+  counter,
+  pages
 })
 
 export default rootReducers
